fix(ci): move ANSI stripping to root conftest, skip auth-dependent tests
- Move ANSI escape code stripping from tests/unit/cli/conftest.py to
  tests/conftest.py so it covers all test directories (unit + integration)
- Mark 4 test classes in test_videos_integration.py with @pytest.mark.auth
  (they require OAuth token file which doesn't exist in CI)
- Exclude auth and slow markers from CI integration test run
- Register auth and slow markers in pytest.ini
