+++ pytest.ini
@@ -14,6 +14,7 @@ markers =
     integration: Integration tests that require external services or full system setup
     docker: marks tests requiring Docker (deselect with '-m "not docker"')
     slow: marks tests as slow (deselect with '-m "not slow"')
+    auth: marks tests requiring OAuth credentials (skipped in CI)
 filterwarnings =
     ignore::pytest.PytestUnraisableExceptionWarning
     ignore::DeprecationWarning:pytest_asyncio*
